# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tekstil Otomasyonu API - OpenAPI 3.0",
        "license": {
            "name": "MIT License",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
        "/personeller": {
            "post": {
                "tags": [
                    "personeller"
                ],
                "summary": "Personel oluşturur.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonelWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Personel bilgileri eksik."
                    },
                    "500": {
                        "description": "Personel Kaydedilemedi."
                    }
                }
            },
            "get": {
                "tags": [
                    "personeller"
                ],
                "summary": "Tüm personelleri getirir.",
                "responses": {
                    "200": {
                        "description": "Personeller Getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Personel"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personel bulunamadı."
                    }
                }
            }
        },
        "/personeller/{id}": {
            "get": {
                "tags": [
                    "personeller"
                ],
                "summary": "ID bilgisine göre personel getirir.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "İstenen personelin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen personel getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personel bulunamadı."
                    }
                }
            },
            "put": {
                "tags": [
                    "personeller"
                ],
                "summary": "Personel günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek personelin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonelWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Personel Güncellendi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Personel Bulunamadı."
                    },
                    "500": {
                        "description": "Personel Güncellenemedi."
                    }
                }
            },
            "delete": {
                "tags": [
                    "personeller"
                ],
                "summary": "Personel siler.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinmek istenen personelin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen personel silindi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personel bulunamadı."
                    },
                    "500": {
                        "description": "Personel silinemedi."
                    }
                }
            },
            "patch": {
                "tags": [
                    "personeller"
                ],
                "summary": "Personel günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek personelin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonelWithoutID"
                            }
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Personel bilgileri eksik."
                    },
                    "500": {
                        "description": "Personel Kaydedilemedi."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Personel": {
                "required": [
                    "personel_id",
                    "ad",
                    "soyad",
                    "ise_baslama_tarih",
                    "unvan",
                    "adres",
                    "sehir",
                    "ulke",
                    "telefon",
                    "dogum_tarihi",
                    "maas",
                    "cinsiyet"
                ],
                "type": "object",
                "properties": {
                    "personel_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 14
                    },
                    "ad": {
                        "type": "string",
                        "example": "Taha"
                    },
                    "soyad": {
                        "type": "string",
                        "example": "Muslu"
                    },
                    "ise_baslama_tarih": {
                        "type": "string",
                        "example": "2020-10-01"
                    },
                    "unvan": {
                        "type": "string",
                        "example": "Yönetici"
                    },
                    "adres": {
                        "type": "string",
                        "example": "Akasya mahallesi 123. sokak NO:12 D:3 Kadıköy"
                    },
                    "sehir": {
                        "type": "string",
                        "example": "İstanbul"
                    },
                    "ulke": {
                        "type": "string",
                        "example": "Türkiye"
                    },
                    "telefon": {
                        "type": "string",
                        "example": "532-123-4567"
                    },
                    "dogum_tarihi": {
                        "type": "string",
                        "example": "1998-10-01"
                    },
                    "maas": {
                        "type": "number",
                        "example": 5000
                    },
                    "cinsiyet": {
                        "type": "string",
                        "example": "E"
                    }
                }
            },
            "PersonelWithoutID": {
                "required": [
                    "ad",
                    "soyad",
                    "ise_baslama_tarih",
                    "unvan",
                    "adres",
                    "sehir",
                    "ulke",
                    "telefon",
                    "dogum_tarihi",
                    "maas",
                    "cinsiyet"
                ],
                "type": "object",
                "properties": {
                    "ad": {
                        "type": "string",
                        "example": "Taha"
                    },
                    "soyad": {
                        "type": "string",
                        "example": "Muslu"
                    },
                    "ise_baslama_tarih": {
                        "type": "string",
                        "example": "2020-10-01"
                    },
                    "unvan": {
                        "type": "string",
                        "example": "Yönetici"
                    },
                    "adres": {
                        "type": "string",
                        "example": "Akasya mahallesi 123. sokak NO:12 D:3 Kadıköy"
                    },
                    "sehir": {
                        "type": "string",
                        "example": "İstanbul"
                    },
                    "ulke": {
                        "type": "string",
                        "example": "Türkiye"
                    },
                    "telefon": {
                        "type": "string",
                        "example": "532-123-4567"
                    },
                    "dogum_tarihi": {
                        "type": "string",
                        "example": "1998-10-01"
                    },
                    "maas": {
                        "type": "number",
                        "example": 5000
                    },
                    "cinsiyet": {
                        "type": "string",
                        "example": "E"
                    }
                }
            },
...
        }
    }
}